extract timer stuff to separate file
